Ajustes dashboard

+++ Sources/_Net/aspnet-core/src/Capitania.Application/DashboardFundo/Dto/DashboardGeralDto.cs
@@ -9,12 +9,12 @@ namespace Capitania.DashboardFundo.Dto
     public class DashboardGeralDto
     {
         public string NomeFundo { get; set; }
-        public float PL { get; set; }
-        public float Caixa { get; set; }
-        public float CRD_EL1 { get; set; }
-        public float VAR { get; set; }
-        public float Stress { get; set; }
-        public float LIQSTRESSPERCENTILE { get; set; }
+        public double PL { get; set; }
+        public double Caixa { get; set; }
+        public double CRD_EL1 { get; set; }
+        public double VAR { get; set; }
+        public double Stress { get; set; }
+        public double LIQSTRESSPERCENTILE { get; set; }
         public DateTime DataInfo { get; set; }
         public DateTime DataRun { get; set; }
     }
